+++ Pandora/Pattern/Functor/Adjoint.hs
@@ -2,7 +2,7 @@ module Pandora.Pattern.Functor.Adjoint where
 
 import Pandora.Pattern.Functor.Covariant (Covariant_)
 
-type (-|) = Adjoint_
+type (-|) = Adjoint
 
 infixl 3 --|-, -|--
 
@@ -14,6 +14,8 @@ infixl 3 --|-, -|--
 > * Right adjunction interchange: psi f ≡ epsilon . comap f
 -}
 
-class (Covariant_ t target source, Covariant_ u source target) => Adjoint_ t u source target where
+-- TODO: Rename Adjoint to Adjoint
+
+class (Covariant_ t target source, Covariant_ u source target) => Adjoint t u source target where
 	(--|-) :: source (t a) b -> target a (u b)
 	(-|--) :: target a (u b) -> source (t a) b