+++ CHANGELOG.md
@@ -495,3 +495,4 @@
 * Define experimental `Distributibe_` typeclass
 * Define experimental `-<<-<<-` method of `Traversable_` typeclass
 * Remove `Adjoint` typeclass
+* Rename experimental `Adjoint_` to `Adjoint`